CP-33338: call vhd-tool with source-format nbdhybrid for NBD sources
This relies on the previous commits that add support for `nbdhybrid` to
`vhd` sparse exports in vhd-tool.

This only affects code paths for VDI.export for all SRs that are
attached through NBD. This means GFS2, and all other SRs currently only
if `prefer-nbd-attach` is set in the config file.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>