ocaml-vhd: Add Hybrid_raw_input to VHD
The function `vhd_from_raw` takes a `input_block` argument, which is a
function that takes a VHD block index and returns a Boolean to indicate
whether the block has data or not. It can be impractical and inefficient
for the caller to provide such a "random access" function, where it may
need to either scan the input multiple times, or keep a large
quick-access data structure in memory.

Instead, `vhd_from_raw` is changes to take a function
`find_data_blocks`, which simply returns a list of the VHD block indices
that have data. These indices are expected to be in accending order,
such that they can be iterated over in an efficient way to construct the
VHD stream.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>