xapi-stdext-encodings: read 4 bytes at a time in ASCII fastpath
Speeds up validation considerably, achieving > 1700 MiB/s.

The idea to use subtraction and bit twiddling for checking the range of
bytes in a word is from:
https://graphics.stanford.edu/~seander/bithacks.html#HasLessInWord

However the formula is simplified significantly taking advantage that we
also check that the MSB in each byte (0x80) is clear in the first place.

Signed-off-by: Edwin Török <edwin.torok@cloud.com>