CA-333441, CA-377454 create /var/lock/sm/iscsiadm
The SM code is supposed to create /var/lock/sm/iscsiadm but it can fail
to do that before we call /opt/xensource/libexec/set-iscsi-initiator. To
reduce the dependency on the SM change, create this directory if
necessary. This facilitates bacporting this change by reducing the
dependency on SM backports to fix this.

Signed-off-by: Christian Lindig <christian.lindig@cloud.com>