Make session errors look less scary in the logs
Currently, if a session check fails and a SESSION_INVALID error is
raised, the message "Session check failed: missing reference; tbl =
session, ref = XX" plus a full strack trace is written to the logs. The
stack trace is usually from session_check.ml and looks the same each
time.

These logs are not very interesting, but unfortunately attact the
attention of people triaging issue by browsing the logs. Because the
SESSION_INVALID lines in xapi's logs look so worrying, they often make
people suspect bugs in xapi. However, the errors just mean that a client
has used an illegal or (more often) expired session. These logs are
especially common when the coordinator xapi is restarted and all old
sessions are automatically removed.

This patch improves logging of session errors by putting the blame where
it belongs: the client (note that clients can be internal). The stack
trace is suppressed, and the log line changed as follows:

    Session check failed: the client used an illegal or expired session
    'X'

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>