swtpm-wrapper: be explicit on when to manufacture a new vTPM
Instead of automatically determining on whether a vTPM needs to be initialized
or not separate this into 2 phases:
* manufacture vTPM if needed
* use vTPM

This will help in implementing the new backend: initialization only works on
local URLs (file:// or dir://).

It also improves error handling: if anything goes wrong during the vTPM
initialization, or initial state upload then we can report back the failure
immediately (instead of an internal QMP error when qemu notices that swtpm
didn't actually start up).

Logs will go to daemon.log, look for 'swtpm-wrapper-<domid>'.

Also unify deprivileged and regular codepaths by always performing a 'chdir()'
into 'tpm_dir', such that relative paths inside and outside the chroot will be
the same.
(otherwise swtpm-wrapper will need to parse and update the URL on whether it is
running inside or outside the chroot which adds complexity when the http
backend is introduced).

TODO: this also switches the state URI to be http, undo that change and further
split this commit up.

Signed-off-by: Edwin Török <edwin.torok@cloud.com>