CP-43387: Fix VDI delta copy with NBD datapath connection
For disks that are backed by NBD rather than a blktap2 device, an update
to sparse_dd is needed to be able to send a VHD difference disk when
copying a VDI with respect to a base VDI. This will also be required for
storage migration.

It is not feasible (efficient) to look at just the NBD exports of the
two VDIs to determine the differences. The proper way to do this is to
add functionality to the SMAPI to obtain the differences between two
VDIs. We will look at this for SMAPIv3 a little later.

In the short term, we'll get sparse_dd to look at the VHD files behind
the scenes. This is how it currently works for the blktap2 devices as
well, where sparse_dd finds the VHD file from the minor number of the
kernel block device.

The NBD server path exposed by tapdisk has its pid and the minor number
encoded in it. This can be extracted and used to find the VHD file as
above.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>