CA-375147: UPDATES_REQUIRE_SYNC when toolstack restarted on coordinator
For repository-based updates, the information of available updates on
every host is stored in a cache (in memory) in XAPI process on
coordinator. This cache aims to make the HTTP GET /updates requests
responded quickly as a fast path. One drawback of this design is that
the cache would be flushed when toolstack on coordinator restarted. As
part of original design, if this happens, an error
'UPDATES_REQUIRE_SYNC' will be raised to clients so that HTTP-based
users would get quick responses in all cases.

This commit aims to fill the cache again (but taking longer time)
automatically instead of raising error 'UPDATES_REQUIRE_SYNC' when the
cache has been flushed. This will simplify the XAPI call flow from
client's perspective.

Signed-off-by: Ming Lu <ming.lu@cloud.com>