fe_test.sh: reliably detect when server is listening on the socket
There is a race condition between 'bind' and 'listen' where the socket exists,
but is not yet accepting connections (this can be tested by artificially
inserting a sleep, or running many times until the race condition is hit like
in the CI).
Create a 2nd file after we called Unix.listen that guarantees that we are now
listening on the socket (we also unlink the file prior to recreating the socket).

Tested by artificially inserting a 'sleep 5' between bind+listen, which
previously caused ECONNECT in the testcase, and after this change it works.

Signed-off-by: Edwin Török <edwin.torok@cloud.com>