CP-42787: Ensure correct pool update sequence
The update of member host should be blocked if pool coordinator is
updated firstly while there is 'RestartToolstack' in coordinator's
recommended guidances, as updates on pool members require new Toolstack
running on coordinator.

Block the update of member host by raising an error:
"require_master_restart_toolstack" in this case.

Signed-off-by: Gang Ji <gang.ji@citrix.com>