CA-354436: pool.is_slave took a long time to respond
This commit is to fix an issue which may be hit when the pool
coordinator is rebooting or its XAPI is restarting.

In either of the cases, other pool members keep trying to connect to
the coordinator for DB connections. This connecting holds the DB lock
until a timer expired. And the timer uses exponential backoff to avoid
overload the coordinator. The maximum of the timer is 256 seconds.

Meanwhile in starting up, the XAPI in coordinator will issue XAPI call
pool.is_slave to every pool member. This call on every pool member will
try to acquire the DB lock as well and hence be blocked by the
re-connecting mentioned above. In the worst case, the start-up of XAPI
would be delayed up to more than 4 minutes (256 seconds).

The idea of this commit is that since the pool.is_slave is issued from
the coordinator, it is reasonable to abort the waiting for the timer
expiry and try connecting again.

Signed-off-by: Ming Lu <ming.lu@cloud.com>