CA-378222: assert_sr_can_host_statefile has to take available space into consideration
If a statefile of the required size is not already present on the SR then
we need to check that there is sufficient space available to create the
statefile.

Otherwise when multiple SRs are available we may pick the wrong one to create
the statefile on, even though a suitable one would've been found if we tried
the other ones.

Signed-off-by: Edwin Török <edwin.torok@cloud.com>