CA-378229: flush database immediately on redo log enable
The redo log protects the XAPI database against (host) crashes.
When the redo log is first enabled there might be some pending data in the XAPI database (in memory),
but this will only be written out when the database is next written to (which could be minutes later).

Perform an immediate database flush when we enable the redo log and force all callers to decide whether they need this
behaviour or not. Code that wants to recover or read back a redo log must not flush the database,
but code that enables the redo log with an already active database must flush it.

This should ensure that when the 'enable_redo_log' API completes we have a crash-proof XAPI database from that point forward.

Signed-off-by: Edwin Török <edwin.torok@cloud.com>