Add HTTP Strict Transport Security header
This header informs the browser that it should use HTTPS.
It is disabled by default. To enable it you need to
specify hsts_max_age that is the number of seconds, after
the reception of the STS header field, during which the
user agent as a Known HSTS Host. Refer to RFC 6797 for
more details about HSTS.

Signed-off-by: Guillaume <guillaume.thouvenin@vates.tech>