CA-378317 fix EBADF in waitpid_nohang (v2, 2/2)
waitpid_nohang closes the socket by calling waitpid and then calls
clear_nonblock on that socket, which fails.

waitpid_nohang's policy around closing the socket is not obvious enough.
Make it more explicit by not calling waitpid but inlining this code and
lookign at the different cases. Also provide an option to never close
the socket such that this resided with the caller - which is a cleaner
protocol. The default behaviour is to still close the socket when the
child process has terminated and not otherwise.

Signed-off-by: Christian Lindig <christian.lindig@cloud.com>