xapi: switch CPU feature sets to the abstract type and don't interpret them
The feature sets are now treated as opaque data in xapi, and operations
on them are delegated to xenopsd. the two main operations are:

1. Combining host feature sets into a pool-level feature set. The
   pool-level represents the lowest common denominator of CPU features.
   It is assigned to VMs that are started on the pool, such that they
   can migrate around to pool, to all hosts. The feature sets used for
   this are the "default policies" from Xen in the
   `features_{hvm;pv}_host` fields.
2. Compatibility checks for live migration. When asked to live-migrate a
   VM to another host, a check is done to ensure that the host has all
   the CPU features that the VM is currently exposed to. This check is
   based on the "max policies" from Xen in the `features_{hvm;pv}`
   fields.

The `Create_misc` module contains the logic for obtaining the hosts
feature sets from xenopsd, and combining them to construct the pool
level. This now uses the new `HOST.combine_cpu_features` call to
xenopsd.

This code used to send alerts for the case that CPU features
were removed from the pool-level, for example when adding a host to the
pool that doesn't have certain features. We can't do this kind of
comparison anymore, and therefore drop the alerting. We may put it back
in another form, later.

The compatibility checks are in `Cpuid_helpers`, which now uses the new
function `HOST.is_compatible`.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>