CP-43565: xapi-expiry-alerts: a new library to generate expiry alerts
This allows to factor out logic for expiration-related alerts to share and
easily test the code.

The interface of this library is "update", upon calling this API, any
existing outdated messages will be removed first, and new message will
be created only if it does not exist in Xapi.Message records yet.

Signed-off-by: Gang Ji <gang.ji@citrix.com>