CA-379173 Implement Stunnel cache using LRU
Replace the Stunndel_cache implementation with a new one, using the LRU
cache module/library.

This seems necessary because the existing implementation is so
convoluted that I have little hope fixing a race condition where the
various hashtables become inconsistent. The new implemenation is
strictly controlled by a lock per cache.

The existing nomenclature is confusing: the cache is really a connection
pool. A connection is taken from the pool while it is used and put back
in after the work is done. The existing limit for the cache size of 70
seems high to me if we are interested in limiting connections.

Signed-off-by: Christian Lindig <christian.lindig@cloud.com>