IH-397: Replace async with lwt in xapi-storage-script
The error handling used in async is translated into Lwt_result type, this is
because the base Lwt only allows exception, which isn't enough to code the
errors produced.

Logging has been changed to use logs with an asynchronous reporter. Now
loglines print the loglevel.

The setup of the loops / promises has been changed, the outer loop has been
removed and its retry-on-error logic has been moved to the two inner watch
loops.

The inotify loop needed quite a few changes as async_inotify was doing complex
handling of events, even consolidating them. This has been simplified by
removing the watches on files and maintaining the directory one, and creating a
list of commands, which the watch loops can act upon, like registering and
deregistering plugins, for example.

Co-developed-by: Edwin Török <edwin.torok@cloud.com>
Signed-off-by: Pau Ruiz Safont <pau.ruizsafont@cloud.com>