CA-377945: toolstack restart: ensure xapi is stopped first, started last
Stopping xapi appears to take a long time for something to time out if
other services have stopped before it. The assumption in
xe-toolstack-restart seems to be that `systemctl stop` with multiple
service names on the same command line respects the `After:` constraints
of the services, but I am not convinced it does. Explicitly stopping
xapi first and starting it after everything else has started, makes the
whole script execute about a minute faster.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>