Add `secureboot-enforce` setting in `xapi.conf`
This setting will be used by varstored to know
wheter to allow the start of a VM that has no certificates
when secureboot is enabled by writing in the xenstore in
`/local/domain/<domid>/platform/secureboot-enforce`.

Default: false to keep the previous behavior.

See: https://github.com/xapi-project/varstored/pull/19

Signed-off-by: BenjiReis <benjamin.reis@vates.fr>