CA-380551: bump minimum HA SR size to 4GiB
Commit 00cb7f2fa changed the default (recommended) size of the redo log to 4GiB,
while keeping the minimum at 256MiB for backwards compatibility with existing VDIs.

However:
* HA must be disabled/re-enabled across updates anyway so backwards compatibility is not needed
* the redo-log on its own is known not to work (bug), keeping backwards compatibility here is not useful
* it makes it more difficult for XenCenter to tell the user when they chose an SR that is too small

If the SR is greater than minimum but below recommended then from the CLI you could still enable HA and create a redo log of the appropriate size,
but not from XenCenter (where the size cannot be chosen).

We could expose a new RO pool field with the minimum size and update clients to use that, but that is unnecessary complication for a very rare use case.

Revert the 'recommended_size' introduction, and instead bump the minimum size.
The existing 'SR.assert_can_host_statefile' can then report an error when the size is wrong and XenCenter can appropriately grey out the choice for the user with the correct message.
(And this method works also for other reasons why we might reject an SR).

Fixes: 00cb7f2fa ("redo-log: bump default size to 4GiB")

Signed-off-by: Edwin Török <edwin.torok@cloud.com>