CA-375992 remove stale swtpm chroots after boot
When a host crashes, chroots in /var/lib/xcp/run remain. Their names
use the domain counter, which resets on reboot. This risks name
collisions between a new chroot and an existing old one. After reboot we
can be sure that no domain is running and that it is safe to remove
stale chroots.

Signed-off-by: Christian Lindig <christian.lindig@cloud.com>