maintenance: prepare mtime usage for 2.0
The conversion from spans to seconds has been removed in this version. In most
cases it's not necessary because we can get away with comparing spans. When
floats representing seconds are needed for time periods, the conversion is done
manually. This is because it's too difficult to change the interface of an API,
be a third-party library of in xapi-idl.

Special care has been given to the scheduler library where it cares only about
the full seconds when converting to span to allow for batching of tasks.

Signed-off-by: Pau Ruiz Safont <pau.ruizsafont@cloud.com>