CA-380178: xenopsd: Fix vTPM manufacture logging
The output from the command is logged afterwards, but this doesn't make
sense because stdout is redirected so it is always empty and stderr is
logged all on one line and is not logged in the case that an exception
occurs (e.g. due to a timeout).

Instead, redirect stderr to syslog as well and skip logging the output.

Signed-off-by: Ross Lagerwall <ross.lagerwall@citrix.com>