forkexecd: handle invalid rpc messages more gracefully
In some situations the string-encoded length is malformed and cannot be
converted into an integer. Use Result.t to relay the cause of the error and
allow callers to log the string that couldn't be decoded. Also shut down the
socket so the other side is forced to stop using the socket when it has
violated the protocol.

Signed-off-by: Pau Ruiz Safont <pau.ruizsafont@cloud.com>