CA-379112 make PBD.plug wait for scan results
For PBD.plug so far, the XAPI DB gets updated as part of an SR scan that
is executed in its own thread. Consequently the XAPI DB may not be up to
date when PBP.plug returns. This leads to a race condition and the
situation observed in CA-379112: the PDB's size is not correct.

The fix is to make the call for the update synchronous.

Signed-off-by: Christian Lindig <christian.lindig@cloud.com>