CA-380789: Not get power_state from snapshots with suspend VDIs
This commit is to fix an issue which was introduced by 3d039f3.
In VM migration, the suspend_vdis will be get from snapshots the
power_state of which is `Suspended. The individual snapshots are
represented as VM records in XAPI as well. They are different with the
base VM record.

This commit reverts this part to the logic prior to 3d039f3.

Signed-off-by: Ming Lu <ming.lu@cloud.com>