CA-380715: Hit unspecified domain_type of snapshots in VM migration
In live migration of a VM with snapshots, the "current_domain_type" of a
snapshot may be "unspecified" if the snapshot was taken when the VM was
in shutdown state. This will fail the CPU checking in migration as the
checking has to know the domain type.

On the other hand, it is not necessary to check CPU for a snapshot if its
domain type is unknown ('unspecified'). The potential problem might be
the snapshot couldn't be reverted and started up in migration destination.
But this can't be avoided anyway. For example, a new created VM which
never started would hit same issue after migration (not live).

This commit skips the CPU checking on such snapshots.

Signed-off-by: Ming Lu <ming.lu@cloud.com>