CA-380551: HA: assert that the HA SR is big enough for BOTH the statefile AND the redo log
Previously only the statefile size was checked, and available space for the database VDI was only checked
during DR in the metadata VDI code, but not during HA in the redolog code.

With this change XenCenter is able to grey out SRs that are too small.

After this change an HA heartbeat SR will have a physical utilization of `3997171712`, which is as close
as we can get to 4GB (the next 4MiB increment would go over 4GB).

Signed-off-by: Edwin Török <edwin.torok@cloud.com>