CA-380581: Remove lock on downloading updates from remote repos
The 'reposync_mutex' was designed for cases in which a pool is
downloading updates from remote repos, but another XAPI call is being
carried on to apply updates simultaneously. The assumption of this
design is that the downloading may take long time. During this period,
it would be nice to allow other operations, e.g get_updates, or
apply_updates, etc.

The issue to be fixed in this commit is that the locked status of
'reposync_mutex' can't be populated to clients since the status of the
lock is not recorded in XAPI DB.

Additionally, with 80c8153860, only the minimum set of update packages
will be downloaded. The assumption above now doesn't hold anymore in
most cases.

The 'pool.allowed_operations' now can take the chance to replace
'reposync_mutex' to protect consistency among repo updates related
operations. As a result, this will populate the locked status to
clients since it is recorded in XAPI DB.

Note: a data change in XAPI DB can be sent/retrieved to/by clients
through API calls.

---------

Signed-off-by: Ming Lu <ming.lu@cloud.com>