CA-381503: bump qemu filesize limit
QEMU has a file size limit to prevent denial of service attacks.
However with the introduction of vTPM the vTPM state is now saved in the qemu device stream.
If the vTPM happens to have a lot of data then the quota is hit, qemu returns an error, the migration fails
and XAPI kills the original VM.
(Same bug happens if we were trying to take a checkpoint, the VM ends up dead).

Fixing that would require fast resume, which we don't have, however the file size limit needs to be updated.
Don't try to be too tight on the size, the vTPM may be a max of 256KiB, so increasing this to 512KiB should work,
but to be safe double that to 1MiB.

Signed-off-by: Edwin Török <edwin.torok@cloud.com>
(cherry picked from commit 177aead9bde0e8ad672b557dc76bfc679ed2f0ed)