CA-381133: Set pending_guidances based on recommended guidance
After applying updates, the `recommended_guidances` field is set to the
"recommended guidance" from the applied metadata from the applied
updates. Any EvacuateHost guidance, which is meant to be followed before
applying updates, is excluded from the list. Xapi uses this field to
just what to do when the client calls `host.apply_recommended_guidance`.
This has not changed.

The `pending_guidances` field used to be set based on the "absolute
guidance" of the update metadata. However, this turned out to be the
wrong choice to guide clients on what to do after applying updates. The
field also contained guidance to reboot the host in case a live-patch
failed to apply. This is not included in the `recommended_guidance` as
it is not meant to be actioned on in `host.apply_recommended_guidance`.

Now, `pending_guidances` is set to the same value as
`recommended_guidances`, plus the reboot guidance in case of a failed
live-patch.

The above applies to the guidance fields of both the `host` and `VM`
classes.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>