[opt] CP-43769: locking_helpers: factor out and reduce allocations in thread local storage
Instead of continously inserting and removing entries from a global integer map
use a hashtable: first time we want to store some per thread data we allocate,
and after that we efficiently find and update the thread's data.
When the thread exits it will be dropped from the hashtable (except the main thread).

Snapshotting becomes less efficient: we need to make a copy of the hashtable, but this is in a slowpath (diagnostic API),
and instead the fastpath avoids allocations.

Ephemerons could've been used, but they currently always allocate an option value internally on lookup.

Signed-off-by: Edwin Török <edwin.torok@cloud.com>