CP-44988: remove API: host.apply_recommended_guidances
This API was added in 23.18.0 for applying all the recommended guidances
after updates are applied. While it is found that the order in which
some recommended guidances are applied is not correct: a
RestartDeviceModel is applied before RestartToolstack, this is a bug:

1. if the new toolstack is already unpacked it may have backward
incompatible scripts vs the running xapi/xenopsd, so you really must
restart the toolstack asap, and not do some migrations first which will
very likely fail completely.

2. if there is a (security) fix in xenopsd in how we launch the device
model (qemu) then obviously that code needs to be active first but we
don't have a way to express that.

And since XC do not use this API to apply recommended guidance(XC calls
corresponding APIs to apply each guidance one by one), this problematic
API is simply removed to resolve the problem.

Signed-off-by: Gang Ji <gang.ji@citrix.com>