CA-381587: log when HTTP Basic auth is used, and by who
For HTTP requests with Basic auth, xapi internally calls
session.login_with_password. The way this is logged makes it look like
the call come from inside the host, especially in case of authentication
errors. This log line makes it explicit what is going on.

The login_with_password call is still logged as coming from xapi itself
(as it is), but the originator argument is used to show where the
external call comes from.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>