CP-42949: Ensure storage RRDs are created without tapdev in kernel
All the necessary data is available also from tapdisk/blktap3 directly,
through the RRD plugin transport protocol. However, this was missed
because of an assumption that tapdev kernel devices are always present.
This is no longer the case.

The code is reorganised to take this into account. In the near future,
we can probably just delete all of the tapdev-dependent code.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>