CP-45175: Enable the OVMF debugging port by default
Currently, getting debug messages out of OVMF requires (amongst other
things) editing the qemu-wrapper script to enable the debug port. This
is inconvenient. Fix this by simply always enabling it.
With the current OVMF, this results in no change in the default
behaviour since the release build of OVMF does not log anything though
that is expected to change in the near future.

Signed-off-by: Ross Lagerwall <ross.lagerwall@citrix.com>