CP-44320 scaffolding for NVidia Virtual Compute Service (VCS)
NVidia is supporting vGPUs that can run as compute vGPUs. These are
marked in their vgpuConfig.xml with a new value "Compute" for the
existing "class" attribute:

<vgpuType id="307" name="GRID P100C-12C" class="Compute">

We don't parse and use this attribute so far but need to in order to
support VCS. This commit extends the parser to recognise the class
attribute and to pass it on under the internal name "vclass" to xenopsd.
Xenopsd's job is to start device emulators accordingly. Selecting the
correct emulators is future work. So far, xenopsd logs the vclass it
receives.

Unit tests have been updated.

This commit should have no impact on current functionality, be safe to
merge, and lay groundwork for extending xenopsd to support VCS.

Signed-off-by: Christian Lindig <christian.lindig@cloud.com>