Publish new Hugo-based docs
This Github Action publishes the new Hugo-based docs from /doc to a temporary
staging area at https://xapi-project.github.io/new-docs/. Eventually, this will
replace the old docs on https://xapi-project.github.io. However, we are still
transferring docs to the new world, which we do incrementally so that it all
can be updated and reviewed again.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>