CP-43755: Locking_helpers: introduce Semaphore
Backport of 8aa69bba142fde6cfffe4b9dbd63dd0e2051a1ba. This required to
introduce a Semaphore_vendored to have access to the Counting
sub-module.

Small wrapper around OCaml's Semaphore module.
It doesn't yet use Thread_state to avoid potential performance issues.

The semaphore is initialized with a count of 1, and its maximum can be
changed at runtime using [set_max] in a safe way that takes into account
threads that already use the semaphore: we store the current max in the
record and call [acquire] or [release] the appropriate number of times
to make it match the new desired max.  This always works, even if there
are threads currently using the semaphore: if we decrease it below max
then 'set_max' will block until sufficient number of threads have
released the semaphore.

[set_max] will be useful for implementing the semaphore maximum value as
a datamodel field, and won't require restarting XAPI to change it.  Even
reading this value from a config file would be difficult otherwise
because the semaphore needs to be initialized with a count at the time
it is created, which is before any configuration files are read.

No functional change.

Signed-off-by: Edwin Török <edwin.torok@cloud.com>
Signed-off-by: Christian Lindig <christian.lindig@cloud.com>