CA-371002: Do a usual import when a default template cannot be found
A usual import is performed when importing a default template if the
template with the same name cannot be found in the database. This allows
the user to export a default template and import it again without
error. When the force option is not enabled, the re-imported template
will no longer be considered default. This enables deletion of such
template.

Signed-off-by: Vincent Liu <shuntian.liu2@cloud.com>