CA-375396: Ignore removed fields when redoing writes
When loading a database from disk in XML, fields that are marked in the
datamodel as removed are ignored and not written to the in-memory database.
Because some fields that have been marked as removed are still written to, like
host metrics' free memory, they might still be tried to be replayed and they
fail because the row does not exist in the database.

The plan is to change the generation of code from the datamodel which will make
the writes impossible. In the meantime follow the same strategy of the xml
loading: do not replay the writes if the field has been marked as removed

Signed-off-by: Pau Ruiz Safont <pau.ruizsafont@cloud.com>