CP-38020: drop numa-placement-strict
This was a debug option that would've by default ignored any exceptions raised by the NUMA planning code.
And when enabled it would've let those exceptions propagate, however it still didn't ensure that the NUMA placement actually worked.

Remove this and treat it as =true, i.e. *if* NUMA placement is enabled then we want any assertions or errors from the code to be raised.
Note that these errors should only be raised if there is a bug in the code: the planner will always find a solution, even if suboptimal
(i.e. place the VM on all NUMA nodes).

No functional change: this doesn't enable NUMA placement by default yet.

Signed-off-by: Edwin Török <edwin.torok@cloud.com>