CP-38020: introduce Host.numa_affinity_policy
The default is still Any, but it can now be set to Best_effort at runtime,
without restarting the toolstack.
This won't change the placement of already running VMs though.

Use an explicit 'default' as the default, so that we can change the default in the future
without having to update the database, but still respect the user's choice if they have explicitly set a certain policy.
And if they haven't explicitly changed this then we'll override the default without changing the value in the DB.

Signed-off-by: Edwin Török <edwin.torok@cloud.com>