CA-365486: repository-domain-name-allowlist could accept a full hostname
For the check of function `assert_url_is_valid`, repository-domain-name-allowlist
doesn't work if you use the full FQDN for a server.

The `assert_url_is_valid` only allows if the host ends with .<entry in repository-domain-name-allowlist>.
Now it is extended to allow if the host matches the entry in the list too.

Signed-off-by: Feiya Zhang <feiya.zhang@cloud.com>