CA-384936 attach static VDIs for redo-log (#5235)
* CA-384936 enable redo log logging

We have a case where we don't know why the redo log is not found on
boot. We need this log entry to understand if any VDIs are found.

Signed-off-by: Christian Lindig <christian.lindig@cloud.com>

* CA-384936 attach static VDIs for redo-log

The redo-log without HA is missing to attach the static VDIs that
actually contain the redo log. Call the same code that we are using in
the HA case.

Signed-off-by: Christian Lindig <christian.lindig@cloud.com>

---------

Signed-off-by: Christian Lindig <christian.lindig@cloud.com>
Co-authored-by: Christian Lindig <christian.lindig@cloud.com>