CP-46264 deprecate host.bios_strings[hp-rombios] entry
The host object maintains a map bios_strings that has an entry for the
"hp-rombios key. The value is based on a value read from dom0 memory
using /dev/mem. This is incompatible with secure boot and this patch
makes the value returned the empty string. The value is patched into a
guest's memory so we can't return anything else.

We might want to remove the key `hp-rembios` in the future.

Signed-off-by: Christian Lindig <christian.lindig@cloud.com>