CP-46140 PVS IPv6 accept both IPv4, IPv6 for PVS Server
We want to accept IPv6 addresses for PVS Server in addition to IPv4
which are currently accepted. This patch widens the predicate that
enforces the format and updates the documentation in the data model.

The address is later written to xenstore where it is picked up by
setup-pvs-proxy-rules which will have to respond to addresses that are
in IPv6 format.

Signed-off-by: Christian Lindig <christian.lindig@cloud.com>