CA-385323: fix xapi-guard reconnection on toolstack restart
On startup xapi-guard will attempt to listen on 'xapi-depriv-socket' for previously running VMs.
It first unlinks the socket, then attempts to connect to XAPI to fetch the VM UUID for each socket, and then call 'bind'. If this fails then the socket will be gone and swtpm/varstored cannot serve those VMs until xapi-guard is restarted again.

This is fine when xapi-guard is restarted on its own (XAPI is up), but it fails during toolstack restart where XAPI may not have completed initializiation by the time xapi-guard attempts to start.

We could fix this by adding a dependency on xapi-init-complete in varstored-guard.service,
but debugging that if it goes wrong is more difficult.
Instead define a new 'reconnect' API which XAPI will call on startup, which gets explicitly logged, and is easier to trace if it goes wrong.

Signed-off-by: Edwin Török <edwin.torok@cloud.com>