CA-365059: Clear source pool messages after migrating VM
Looks like we weren't moving messages if the VM was on a supporter at all in the past.
This fixes both the issue in the ticket and moves VM messages if the VM was originally on a coordinator.
Using RPC calls since internal xapi calls fail, see #5216 and its revert PR #5239.

Signed-off-by: Danilo Del Busso <danilo.delbusso@cloud.com>